# ENH: add masked algorithm for mean() function