--- a/doc/source/whatsnew/v1.3.0.rst
+++ b/doc/source/whatsnew/v1.3.0.rst
@@ -160,7 +160,7 @@ Deprecations
 Performance improvements
 ~~~~~~~~~~~~~~~~~~~~~~~~
 - Performance improvement in :meth:`IntervalIndex.isin` (:issue:`38353`)
-- Performance improvement in :meth:`Series.mean` for ``ExtensionDtype`` columns (:issue:`34814`)
+- Performance improvement in :meth:`Series.mean` for nullable data types (:issue:`34814`)
 -
 
 .. ---------------------------------------------------------------------------