--- a/pandas/core/array_algos/masked_reductions.py
+++ b/pandas/core/array_algos/masked_reductions.py
@@ -110,5 +110,8 @@ def max(values: np.ndarray, mask: np.ndarray, *, skipna: bool = True):
 
 
 def mean(values: np.ndarray, mask: np.ndarray, skipna: bool = True, min_count: int = 0):
-    return sum(values=values, mask=mask, skipna=skipna, min_count=min_count) / np.count_nonzero(~mask)
+    return _sumprod(
+        np.sum, values=values, mask=mask, skipna=skipna, min_count=min_count
+    ) / np.count_nonzero(~mask)
+