--- a/pandas/core/array_algos/masked_reductions.py
+++ b/pandas/core/array_algos/masked_reductions.py
@@ -113,5 +113,3 @@ def mean(values: np.ndarray, mask: np.ndarray, skipna: bool = True, min_count: i
     return _sumprod(
         np.sum, values=values, mask=mask, skipna=skipna, min_count=min_count
     ) / np.count_nonzero(~mask)
-
-