--- a/pandas/core/array_algos/masked_reductions.py
+++ b/pandas/core/array_algos/masked_reductions.py
@@ -110,7 +110,8 @@ def max(values: np.ndarray, mask: np.ndarray, *, skipna: bool = True):
 
 
 def mean(values: np.ndarray, mask: np.ndarray, skipna: bool = True):
-    
+    if not values.size:
+        return libmissing.NA
     _sum = _sumprod(np.sum, values=values, mask=mask, skipna=skipna)
     count = np.count_nonzero(~mask)
     mean_value = _sum / count