@pep8speaks 2020-06-16
Hello @Akshatt! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`pandas/core/array_algos/masked_reductions.py`](https://github.com/pandas-dev/pandas/blob/bcf8f058008feb7f9c97264f9fd2c9c26746eed8/pandas/core/array_algos/masked_reductions.py):

> [Line 106:6](https://github.com/pandas-dev/pandas/blob/bcf8f058008feb7f9c97264f9fd2c9c26746eed8/pandas/core/array_algos/masked_reductions.py#L106): [E111](https://duckduckgo.com/?q=pep8%20E111) indentation is not a multiple of four
> [Line 106:6](https://github.com/pandas-dev/pandas/blob/bcf8f058008feb7f9c97264f9fd2c9c26746eed8/pandas/core/array_algos/masked_reductions.py#L106): [E117](https://duckduckgo.com/?q=pep8%20E117) over-indented
> [Line 106:105](https://github.com/pandas-dev/pandas/blob/bcf8f058008feb7f9c97264f9fd2c9c26746eed8/pandas/core/array_algos/masked_reductions.py#L106): [W292](https://duckduckgo.com/?q=pep8%20W292) no newline at end of file