@jorisvandenbossche 2020-06-16
See here:

https://github.com/pandas-dev/pandas/blob/4628665e53c214ae2e2efde54c4d703fb1c198ab/pandas/core/arrays/masked.py#L337-L339

You can add "mean" there