@Akshatt 2020-06-16
@jorisvandenbossche, 
The new mean function takes significantly less time 
![Screenshot (189)](https://user-images.githubusercontent.com/32221186/84778979-60942980-b001-11ea-982b-1d9f32d7c075.png)

Should tests be added? and if yes then where?
and same for whatsnew entry