@jreback 2020-06-16
can you add some asv's for the existing and this new case (perf metrics)