@Akshatt 2020-06-17
@jreback 
> pls add tests (we likely aready have some setup for sum, you can just add onto those)

Alright, so I need to add tests in 
* [test_reductions](https://github.com/pandas-dev/pandas/blob/5fdd6f50ac5ac2de939ef5b849e4c41e27e623a3/pandas/tests/reductions/test_reductions.py#L565)   
* [test_dtypes](https://github.com/pandas-dev/pandas/blob/5fdd6f50ac5ac2de939ef5b849e4c41e27e623a3/pandas/tests/arrays/integer/test_dtypes.py#L22)  
Correct me if im wrong or need to add in more tests.

> can you add some asv's for the existing and this new case (perf metrics)

Where exactly do I add the asv's in?