@jorisvandenbossche 2020-06-17
This operation is already covered by

https://github.com/pandas-dev/pandas/blob/107ad154c65697be0f88d6519a9c5abe8fe29d70/asv_bench/benchmarks/series_methods.py#L250-L280

(I added Int64 in general for all ops, so including mean, when I added a masked sum/prod algorithm)

So no need to write additional benchmarks I think.

For tests, it would be good to add it here: https://github.com/pandas-dev/pandas/blob/5fdd6f50ac5ac2de939ef5b849e4c41e27e623a3/pandas/tests/reductions/test_reductions.py#L683 (we mainly need to check the behaviour on empty)