@Akshatt 2020-06-17
@jorisvandenbossche, 
> We are also going to need to decide what to return for an empty array/Series. Right now we return np.nan, but this can also be pd.NA. This touches a bit the discussion we had about NaN vs NA for the floating dtype, but I suppose for now we probably want to go with pd.NA (that's what we do now).

This would involve 
* checking if the mean is np.nan and then  
* returning pd.NA instead in the mean function itself  
yes?