@dsaxton 2020-09-15
> This would involve
> 
> * checking if the mean is np.nan and then
> * returning pd.NA instead in the mean function itself
>   yes?

@Akshatt You could also do a length check before any calculation as with other functions here, e.g. something like `not values.size`