@dsaxton 2020-09-15
@Akshatt  Since we're always returning NA on empty input regardless of skipna could also
```python
if not values.size:
    return pd.NA

# logic for non-empty input
```