@Akshatt 2020-12-13
@jorisvandenbossche 
Looks like the tests are not passing for `float64` and `object` dtypes. They return `nan` for empty series and not `pd.NA`. Should I keep a check for them separately or remove them from the parameterized dtypes?