@jorisvandenbossche 2020-12-13
Yes, the "object" dtype can be removed (it's not a nullable dtype), and "float64" can be changed to "Float64", see my inline comments