@jreback 2020-12-14
do we have an asv that hits this? also need a whatsnew note for 1.3 that says perf increased for mean on an EA masked array.