@jorisvandenbossche 2020-12-14
> do we have an asv that hits this?

Yes, `stat_ops.py::FrameOps` and `series_methods.py::NanOps` already cover this