@jreback 2020-12-23
can you check if we have asv's as indicated inthe OP and show the results of them. (as this was the point of adding a specific function here)