@jreback 2020-12-23
> > adding min_count ?
> 
> As I already answered, `min_count` is not applicable here. So I think it is only the whatsnew note that remains code-wise

this is fine (you are repeating an old commment which i agree is not relevant).

the asv's and whatsnew are still needed.